Move TestUtil into util package
