+++ src/test/scala/org/hammerlab/guacamole/LociSetSuite.scala
@@ -21,7 +21,7 @@ package org.hammerlab.guacamole
 import org.apache.commons.io.FileUtils
 import org.apache.hadoop.fs.FileUtil
 import org.apache.spark.rdd.RDD
-import org.hammerlab.guacamole.util.GuacFunSuite
+import org.hammerlab.guacamole.util.{ TestUtil, GuacFunSuite }
 import org.hammerlab.guacamole.reads.Read
 import org.scalatest.Matchers
