+++ src/test/scala/org/hammerlab/guacamole/reads/ReadSetSuite.scala
@@ -21,8 +21,7 @@ package org.hammerlab.guacamole.reads
 import org.apache.spark.rdd.RDD
 import org.bdgenomics.adam.rdd.ADAMContext
 import org.bdgenomics.formats.avro.AlignmentRecord
-import org.hammerlab.guacamole.util.GuacFunSuite
-import org.hammerlab.guacamole.TestUtil
+import org.hammerlab.guacamole.util.{ TestUtil, GuacFunSuite }
 import org.scalatest.Matchers
 import org.bdgenomics.adam.rdd.ADAMContext._
