+++ src/test/scala/org/hammerlab/guacamole/reads/UnmappedReadSerializerSuite.scala
@@ -18,9 +18,8 @@
 
 package org.hammerlab.guacamole.reads
 
-import org.hammerlab.guacamole.util.GuacFunSuite
-import org.hammerlab.guacamole.TestUtil
-import org.hammerlab.guacamole.TestUtil.Implicits._
+import org.hammerlab.guacamole.util.{ TestUtil, GuacFunSuite }
+import TestUtil.Implicits._
 import org.scalatest.Matchers
 
 class UnmappedReadSerializerSuite extends GuacFunSuite with Matchers {
